// generated/src/app/api-models/auth-addtl-info-pwk.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AuthAddtlInfoPwk {
  seqAuthAddtlInfoPwk: number;
  authNumber: number;
  secondaryAuthNo: string;
  seqAuthProcId: number;
  respReqFlg: string;
  processStatus: string;
  rptTypeCode: string;
  transCode: string;
  ctrlNo: string;
  description: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}